requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
python-multipart==0.0.22
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic==2.10.3
pydantic[email]
pydantic-settings==2.6.1
python-dotenv==1.1.0

langchain-core
langchain-gigachat
langgraph
langfuse

pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.2
pdfplumber==0.11.9
pymupdf==1.25.5
python-docx==1.1.2
docx2pdf==0.1.8
Pillow==11.1.0
openai==2.24.0
httpx==0.28.1
email-validator==2.3.0
